requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.8.0
Django==5.2.3
django-bootstrap5==25.1
django-decouple==2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
fonttools==4.59.0
idna==3.10
inflection==0.5.1
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
numpy==1.24.3
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==5.2.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
scikit-surprise==1.1.4
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
gunicorn
psycopg2>=2.9.9
whitenoise==6.6.0
django-jazzmin
pytest
pytest-django
flake8
black